backend/requirements.txt
Django==1.11.7
uwsgi==2.0.17
python-dotenv==0.7.1
pytz==2017.3
djangorestframework==3.7.7
django-cors-headers
psycopg2==2.7.3.2
djangorestframework-recursive==0.1.2